Add asset version number to cache bursting query string
Assets are burst out of Concrete5's own cache based on changing the version number.
The same should be true for the browser's cache. Relying solely on the package's version number is not enough. As an example, in my package Buttons Factory, I generate a CSS file and the asset version number is set dynamically.

This adds the version number, if any, to the string before obfuscating it.